requirements.txt
# SmartNotes AI - Requirements (Text, PDF, URL Support Only)
# Python 3.11.0 Compatible

# ==================== CORE PACKAGES ====================
Flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
jinja2==3.1.2

# ==================== AI/ML PACKAGES ====================
transformers==4.33.2
torch==2.0.1
nltk==3.8.1
numpy==1.24.3

# ==================== FILE PROCESSING ====================
PyPDF2==3.0.1
pdfplumber==0.9.0
reportlab==4.0.4
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.21

# ==================== LANGUAGE SUPPORT ====================
langdetect==1.0.9
googletrans==4.0.0rc1

# ==================== AUTHENTICATION & DATABASE ====================
flask-login==0.6.2
flask-sqlalchemy==3.0.5
email-validator==2.0.0

# ==================== URL/WEBSITE SCRAPING ====================
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# ==================== UTILITIES ====================
python-dotenv==1.0.0
pillow==10.0.1